dom/bindings/test/test_exceptionSanitization.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1295322
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1295322</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1295322">Mozilla Bug 1295322</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  /** Test for Bug 1295322 **/
  iframe = document.createElement('iframe');
  iframe.name = "eWin";
  document.body.appendChild(iframe);

  try{
    // NOTE: The idea here is to call something that will end up throwing an
    // exception in a JS component and then propagate back to C++ code before
    // returning to us.  If opening a feed: URI stops doing that, we will need a
    // new guinea pig here.
    open('feed://java:script:codeshouldgohere','eWin');
    ok(false, "Should have thrown!");
  } catch(e){
    try {
      is(e.name, "NS_ERROR_UNEXPECTED", "Should have the right exception");
      is(e.filename, location.href,
         "Should not be seeing where the exception really came from");
    } catch (e2) {
      ok(false, "Should be able to work with the exception");
    }
  }
  </script>
</body>
</html>
